refactor: separate sheet and spreadsheet logic
