+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Google/Tools/Comparers/ISheetTitleComparer.cs
@@ -1,0 +1,6 @@
+﻿namespace Kysect.Shreks.Integration.Google.Tools.Comparers;
+
+public interface ISheetTitleComparer : IComparer<string>
+{
+
+}
