+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Google/Tools/ISheetManagementService.cs
@@ -1,4 +1,6 @@
-﻿namespace Kysect.Shreks.Integration.Google.Tools;
+﻿using Google.Apis.Sheets.v4.Data;
+
+namespace Kysect.Shreks.Integration.Google.Tools;
 
 public interface ISheetManagementService
 {
@@ -9,12 +11,8 @@ public interface ISheetManagementService
     /// <returns>Sheet id</returns>
     Task<int> CreateOrClearSheetAsync(string spreadsheetId, string sheetTitle, CancellationToken token);
 
-    /// <summary>
-    ///     Creates spreadsheet in drive,
-    ///     configured by DriveParentProvider
-    /// </summary>
-    /// <returns>Spreadsheet id</returns>
-    Task<string> CreateSpreadsheetAsync(string title, CancellationToken token);
+    /// <returns>Sheet id</returns>
+    Task<int> CreateSheetAsync(string spreadsheetId, string sheetTitle, CancellationToken token);
 
     Task<bool> CheckIfExists(string spreadsheetId, string sheetTitle, CancellationToken token);
 }
