--- Source/Infrastructure/Integration/Kysect.Shreks.Integration.Google/Tools/ISheetTitleComparer.cs
@@ -1,5 +1,0 @@
-﻿namespace Kysect.Shreks.Integration.Google.Tools;
-
-public interface ISheetTitleComparer : IComparer<string>
-{
-}
